2 этап основной сервис, main_svc остальные эндпоинты интеграция с сервисом статистики.

+++ stats/statsDto/src/main/java/ru/practicum/dtos/ViewStatsDto.java
@@ -10,5 +10,5 @@ import lombok.*;
 public class ViewStatsDto {
     private String app;
     private String uri;
-    private Long hits;
+    private int hits;
 }
